requirements.txt
aiohappyeyeballs~=2.4
aiohttp~=3.10
aiosignal~=1.3
annotated-types~=0.7
anyio~=4.6
appnope~=0.1.4
argon2-cffi~=23.1
argon2-cffi-bindings~=21.2
arrow~=1.3
asgiref~=3.8
asttokens~=2.4
async-lru~=2.0
attrs~=24.2
babel~=2.16
backoff~=2.2
bcrypt~=4.2
beautifulsoup4~=4.12
bleach~=6.2
build~=1.2
cachetools~=5.5
certifi~=2024.8
cffi~=1.17
charset-normalizer~=3.4
chroma-hnswlib~=0.7
chromadb~=0.5
click~=8.1
coloredlogs~=15.0
comm~=0.2
dataclasses-json~=0.6
debugpy~=1.8
decorator~=5.1
defusedxml~=0.7
Deprecated~=1.2
durationpy~=0.9
executing~=2.1
fastapi~=0.115
fastjsonschema~=2.20
filelock~=3.16
flatbuffers~=24.3
fqdn~=1.5
frozenlist~=1.5
fsspec~=2024.10
google-ai-generativelanguage~=0.6
google-api-core~=2.22
google-api-python-client~=2.151
google-auth~=2.36
google-auth-httplib2~=0.2
google-generativeai~=0.8
googleapis-common-protos~=1.65
greenlet~=3.1
grpcio~=1.67
grpcio-status~=1.67
h11~=0.14
httpcore~=1.0
httplib2~=0.22
httptools~=0.6
httpx~=0.27
httpx-sse~=0.4
huggingface-hub~=0.26
humanfriendly~=10.0
idna~=3.10
importlib_metadata~=8.5
importlib_resources~=6.4
ipykernel~=6.29
ipython~=8.29
isoduration~=20.11
jedi~=0.19
Jinja2~=3.1
json5~=0.9
jsonpatch~=1.33
jsonpointer~=3.0
jsonschema~=4.23
jsonschema-specifications~=2024.10
jupyter-events~=0.10
jupyter-lsp~=2.2
jupyter_client~=8.6
jupyter_core~=5.7
jupyter_server~=2.14
jupyter_server_terminals~=0.5
jupyterlab~=4.2
jupyterlab_pygments~=0.3
jupyterlab_server~=2.27
kubernetes~=31.0
langchain~=0.3
langchain-community~=0.3
langchain-core~=0.3
langchain-google-genai~=2.0
langchain-text-splitters~=0.3
langsmith~=0.1
markdown-it-py~=3.0
MarkupSafe~=3.0
marshmallow~=3.23
matplotlib-inline~=0.1
mdurl~=0.1
mistune~=3.0
mmh3~=5.0
monotonic~=1.6
mpmath~=1.3
multidict~=6.1
mypy-extensions~=1.0
nbclient~=0.10
nbconvert~=7.16
nbformat~=5.10
nest-asyncio~=1.6
notebook~=7.2
notebook_shim~=0.2
numpy~=1.26
oauthlib~=3.2
onnxruntime~=1.20
opentelemetry-api~=1.28
opentelemetry-exporter-otlp-proto-common~=1.28
opentelemetry-exporter-otlp-proto-grpc~=1.28
opentelemetry-instrumentation~=0.49
opentelemetry-instrumentation-asgi~=0.49
opentelemetry-instrumentation-fastapi~=0.49
opentelemetry-proto~=1.28
opentelemetry-sdk~=1.28
opentelemetry-semantic-conventions~=0.49
opentelemetry-util-http~=0.49
orjson~=3.10
overrides~=7.7
packaging~=24.1
pandas~=2.2
pandocfilters~=1.5
parso~=0.8
pexpect~=4.9
platformdirs~=4.3
posthog~=3.7
prometheus_client~=0.21
prompt_toolkit~=3.0
propcache~=0.2
proto-plus~=1.25
protobuf~=5.28
psutil~=6.1
ptyprocess~=0.7
pure_eval~=0.2
pyasn1~=0.6
pyasn1_modules~=0.4
pycparser~=2.22
pydantic~=2.9
pydantic-settings~=2.6
pydantic_core~=2.23
Pygments~=2.18
pyparsing~=3.2
PyPika~=0.48
pyproject_hooks~=1.2
python-dateutil~=2.9
python-dotenv~=1.0
python-json-logger~=2.0
pytz~=2024.2
PyYAML~=6.0
pyzmq~=26.2
referencing~=0.35
requests~=2.32
requests-oauthlib~=2.0
requests-toolbelt~=1.0
rfc3339-validator~=0.1
rfc3986-validator~=0.1
rich~=13.9
rpds-py~=0.21
rsa~=4.9
Send2Trash~=1.8
shellingham~=1.5
six~=1.16
sniffio~=1.3
soupsieve~=2.6
SQLAlchemy~=2.0
stack-data~=0.6
starlette~=0.41
sympy~=1.13
tenacity~=9.0
terminado~=0.18
tinycss2~=1.4
tokenizers~=0.20
tornado~=6.4
tqdm~=4.67
traitlets~=5.14
typer~=0.12
types-python-dateutil~=2.9
typing-inspect~=0.9
typing_extensions~=4.12
tzdata~=2024.2
uri-template~=1.3
uritemplate~=4.1
urllib3~=2.2
uvicorn~=0.32
uvloop~=0.21
watchfiles~=0.24
wcwidth~=0.2
webcolors~=24.8
webencodings~=0.5
websocket-client~=1.8
websockets~=13.1
wrapt~=1.16
yarl~=1.17
zipp~=3.20
